requirements.txt
# Runtime dependencies for Parametric Computer Book Factory
pyyaml>=6.0
python-docx>=1.1.0
qrcode[pil]>=7.4
pillow>=10.0
